+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+- Remove `interactiveCriticalDisabled` color variant, use `textDisabled` instead ([#116](https://github.com/Shopify/polaris-tokens/pull/116))
 
 ## [2.9.0] - 2020-03-03